cambio a mochila 0/1 falta impresion de mochila pd

+++ Labo2_TP1/Labo2_TP1/cQueso.cs
@@ -11,8 +11,8 @@ namespace Labo2_TP1
         //todo VER COMO ORGANIZAR EL CODIGO
 
         public int codigo { get; set; }
-        public float precio { get; set; }
-        public float peso { get; set; }
+        public int precio { get; set; }
+        public int peso { get; set; }
 
     }
 }
